requirements.txt
pylint
requests
websocket-client>=1.6.4
urllib3<2.0.0
